refactor(world): state перенесён в отдельную папку
