+++ packages/world/state/__tests__/mocks.ts
@@ -1,0 +1,18 @@
+import { peetData } from '../../character/__tests__/mocks';
+
+import type { CharacterSlice, WorldSlice, State } from '../types';
+
+export const mockedCharacterState: CharacterSlice = {
+  location: peetData.spawn,
+  data: peetData,
+};
+
+export const mockedWorldState: WorldSlice = {
+  time: { hours: 8, minutes: 2 },
+  day: 1,
+};
+
+export const mockedState: State = {
+  character: mockedCharacterState,
+  world: mockedWorldState,
+};
