+++ packages/world/state/index.ts
@@ -1,0 +1,1 @@
+export type { CharacterSlice, State, WorldSlice } from './types';
