+++ packages/world/tasks/live-day-task.ts
@@ -1,8 +1,8 @@
 import type { ICompoundTask } from '@sentient-world/htn';
 
-import type { WorldState } from '../world/state';
+import type { State } from '../state';
 
-export class LiveDayTask implements ICompoundTask<WorldState> {
+export class LiveDayTask implements ICompoundTask<State> {
   name = 'LiveDayTask';
 
   getMethods() {
