+++ packages/world/tasks/primitives/sleep-task.ts
@@ -1,7 +1,7 @@
 import type { IGeometry, Time } from '@sentient-world/engine';
 import type { ExecutionStatus, IPrimitiveTask } from '@sentient-world/htn';
 
-import type { State } from '../../world';
+import type { State } from '../../state';
 
 export type SleepTaskOptions = {
   until: Time;
