--- packages/world/world/index.ts
@@ -1,1 +1,0 @@
-export type { CharacterState, State, WorldState } from './state';
